Removed old files
